api/requirements.txt
tensorflow==2.15.0
numpy==1.26.2
matplotlib==3.8.2
fastapi==0.104.1
pillow==10.1.0
tensorflow-serving-api==2.14.1
uvicorn==0.24.0.post1
python-multipart==0.0.6
pillow==10.1.0
typing